test(voice-call): split playback-mark coverage into its own suite
